requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.7
django-allauth==65.5.0
gunicorn==23.0.0
idna==3.10
joblib==1.4.2
numpy==2.2.3
packaging==24.2
pandas==2.2.3
pycparser==2.22
pycryptodome==3.21.0
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
requests==2.32.3
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.5.0
tzdata==2025.1
urllib3==2.3.0
